+++ feature/patients/src/main/java/ru/dekabrsky/scenarios/presentation/model/PatientInvitationUiModel.kt
@@ -3,6 +3,6 @@ package ru.dekabrsky.scenarios.presentation.model
 class PatientInvitationUiModel(
     var patientName: String = "",
     var parentName: String = "",
-    var parentId: Int? = null,
+    var parentId: Long? = null,
     var older15: Boolean = false
 )
