+++ feature/patients/src/main/java/ru/dekabrsky/scenarios/presentation/model/SelectParentArgs.kt
@@ -4,6 +4,6 @@ import ru.dekabrsky.common.domain.model.ContactEntity
 import java.io.Serializable
 
 class SelectParentArgs(
-    val selectedParentId: Int? = null,
+    val selectedParentId: Long? = null,
     val variants: List<ContactEntity>
 ): Serializable
